+++ nullpomino-core/src/main/kotlin/mu/nu/nullpo/gui/common/BaseFontNormal.kt
@@ -64,8 +64,8 @@ abstract class BaseFontNormal:BaseFont {
 		}
 	}
 
-	override fun printFont(x:Int, y:Int, str:String, color:COLOR, scale:Float, alpha:Float, rainbow:Int) =
-		processTxt(x.toFloat(), y.toFloat(), str, color, scale, alpha, rainbow)
+	override fun printFont(x:Float, y:Float, str:String, color:COLOR, scale:Float, alpha:Float, rainbow:Int) =
+		processTxt(x, y, str, color, scale, alpha, rainbow)
 		{i:Int, dx:Float, dy:Float, s:Float, sx:Int, sy:Int, sw:Int, sh:Int, a:Float ->
 			getImg(i).draw(dx, dy, dx+sw*s, dy+sh*s, sx, sy, sx+sw, sy+sh, alpha = a)
 		}
